chore: rename BaseFooter to UIFooter
